health_rag_agent/requirements.txt
langchain-community
langchain-openai
pymilvus
fastapi
uvicorn
reportlab
pypdf
unstructured
python-docx
